+++ titanic_problem_set/plot_util.py
@@ -16,6 +16,7 @@ def plot_count(df: pd.DataFrame,
             hue=hue,
             palette=colour_list)
         plt.title(f"{title} / {col}")
+        plt.grid(color="black", linestyle="-.", linewidth=0.5, axis="y", which="major")
     else:
         # set number of columns
         ncols = 2
@@ -37,6 +38,7 @@ def plot_count(df: pd.DataFrame,
                     ax=ax)
                 # Set title and show plot
                 ax.set_title(f"{title} / {col}")
+                ax.grid(color="black", linestyle="-.", linewidth=0.5, axis="y", which="major")
 
         fig.tight_layout()
     plt.show()
@@ -44,18 +46,16 @@ def plot_count(df: pd.DataFrame,
 
 def plot_distribution(df: pd.DataFrame, 
                       features: Union[str, List[str]], 
-                      multiple: str,
                       hue: str,
                       title: str,
-                      colour_list: List) -> None:
+                      color_list: List) -> None:
     if type(features) is str:
         col = features
         sns.histplot(
             data = df.dropna(subset=[col]),  #handle missing values
             x=col,
-            multiple=multiple,
-            hue=hue,
-            palette=colour_list)
+            hue=hue, 
+            palette=color_list)
         plt.title(f"{title} / {col}")
     else:
         # set number of columns (use 3 to demonstrate the change)
@@ -73,11 +73,11 @@ def plot_distribution(df: pd.DataFrame,
                 sns.histplot(
                     data = df.dropna(subset=[col]),  #handle missing values
                     x=col,  
-                    multiple=multiple,
                     hue=hue,
-                    palette=colour_list,
+                    palette=color_list,
                     ax=ax)
             # Set title and show plot
             ax.set_title(f"{title} / {col}")
         fig.tight_layout()
+
     plt.show()
